Bugfix: Implement dedicated SendQueue and InterestQueue with no shared_ptrs
This fixes a very odd bug which crashes the whole system randomly.
When adding and entry to the Queue (template version) by calling pushBack,
the application segfaults on shared_ptr assignment in the function
__gnu_cxx::__exchange_and_add(int volatile*, int).

Google doesn't really come up with anything useful related to that. It
might be a bug in code generation (found an old report that uncovered
such a bug a few years ago) which is related to multithreaded usage.

The best and fastest fix for this was to simply use raw pointers.