Major refactoring to adapt to new and optimized buffer handling
The msg namespace is considered dead (still available here)
because all interest handling was moved up into the java space
to get rid of handling and dispatching of two queues.

A new ring buffer based structure handles serialization by
executing a single copy operation to a ready to send buffer.
This buffer is passed down to the ib subsystem and copied
(once more) to a memory mapped buffer.