ibnet/msgrc: Major refactoring
Move everything related to RC messaging from the former dx
namespace to msgrc. Better separation of IB related data and logic.
More abstraction to create a loopback test and new JNI binding
for dxnet. A lot of cleanup and performance improvements.

The whole send/recv logic is re-newed to ensure better utilization
of the hardware