msgrc/SendDispatcher: Add counters to detect rnr naks
There is no possible to check for rnr naks and how long they
forced the sending HCA to wait. Use a timer to detect if the sender
has added something to the send queue and has not sent it for at
least 100ms. Use three counters to differ > 100ms, > 250ms and
> 500ms delays